+++ packages/graph-explorer/src/core/StateProvider/index.ts
@@ -1,5 +1,6 @@
 export * from "./appStore";
 export * from "./configuration";
+export * from "./defaultStylingAtom";
 export * from "./displayAttribute";
 export * from "./displayEdge";
 export * from "./displayTypeConfigs";